♻️ 哲学者のアクションを別ファイルに分けた
