+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+	"files.associations": {
+		"array": "c",
+		"string": "c",
+		"string_view": "c",
+		"vector": "c"
+	}
+}
