+++ philo/Makefile
@@ -6,7 +6,7 @@
 #    By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/08 01:21:55 by miyuu             #+#    #+#              #
-#    Updated: 2025/04/21 21:35:59 by miyuu            ###   ########.fr        #
+#    Updated: 2025/04/21 22:24:03 by miyuu            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -32,7 +32,8 @@ SRC_FILES = main.c \
 			ft_atoi.c \
 			ft_isdigit.c \
 			utils.c \
-			take_put_forks.c
+			take_put_forks.c \
+			action_philosophers.c
 
 # Header files
 HEADER = $(HEADER_DIR)/philo.h
