+++ philo/include/philo.h
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 19:53:30 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/21 21:35:08 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/21 22:23:56 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,7 +94,7 @@ long			get_now_time_ms(void);
 void			*judgement_stop_thread(void *arg);
 void			take_forks(t_thread_arg *philo, t_univ_rules rules);
 void			put_forks(t_thread_arg *philo);
-
+void			*action_philosophers(void *arg);
 int				ft_atoi(const char *str);
 int				ft_isdigit(int c);
 void			safe_usleep(int time_ms);
