+++ philo/src/utils.c
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 21:31:57 by miyuu             #+#    #+#             */
-/*   Updated: 2025/04/21 21:32:13 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/04/21 22:22:57 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
